Unify ApplicationUI trait from components
While it seemed more composable at the time, I was spending more time
attempting to shoehorn the user interface into the specified components
instead of handling the events more opaquely. The unified interface
seems clearer now though it adds a bit of indirection to ensure the
cocoa implementation of ApplicationUI is safe.